PWM Generation code using timer 1
In this program, the generation of a PWM signal is made by using timer 1. There are 4 methods: 2 of them are used to make a linear approximation, the third one is used to map the values of the joystick to the values in milliseconds (duty-cycle) and the final one is used to update the CRR value of a channel of the timer 1.